Add task #5
The designer made a template for home rugs. For the mass production of rugs, he needs to be able to quickly make a layout of an arbitrary size. It is known that the length of the rug is always 3 times greater than its width (W = 3 * H).

Given: the width of the rug.

Assignment: write a program that will make a rug layout for its further production.