Add task #1
Given: A number, as an integer (int).
Exercise:
"Fizz buzz" is a word game with which we will teach our robots how to divide. Let's train the computer.
You must write a program that takes a positive integer and returns the sl. values.

"Fizz Buzz" if the number is divisible by 3 and 5;
"Fizz" if the number is divisible by 3;
"Buzz" if the number is divisible by 5;
Number, as a string for other cases.

Example:

  x = 15, result: "Fizz Buzz"
  x = 6, result: "Fizz"
  x = 5, result: "Buzz"
  x = 7, result: "7"