Add task #6
Given: a sequence of lines.
Task: You must combine these lines into a block of text, separating the original lines with commas. you must replace all occurrences of the word "right" with the word "left", even if it is part of another word. All lines are in lower case.

Examples:
["left", "right", "left", "stop"], result: "left,left,left,stop"
["bright aright", "ok"], result: "bleft aleft,ok"
["enough", "jokes"], result: "enough,jokes"